new structure with diffusion2d directory, added pyproject.toml
